move: child폴더 내 컴포넌트들 _components폴더에서components폴더로 이동
